requirements.txt
aio-pika==9.5.5
aiofiles==24.1.0
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.6.2
async-property==0.2.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorlog==6.9.0
coverage==7.8.0
cryptography==44.0.1
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
iniconfig==2.1.0
jwcrypto==1.5.6
motor==3.7.0
multidict==6.2.0
packaging==24.2
pamqp==3.3.0
pillow==11.1.0
pip==25.0
pluggy==1.5.0
propcache==0.3.0
psycopg2==2.9.9
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.2
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
python-jose==3.4.0
python-json-logger==3.2.1
python-keycloak==5.3.1
python-magic==0.4.27
python-multipart==0.0.20
qrcode==8.0
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==76.0.0
sniffio==1.3.0
SQLAlchemy==2.0.37
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wheel==0.45.1
yarl==1.18.3
